+++ R/generics.R
@@ -54,7 +54,6 @@
 #'
 #' * `mutate()` generates a list of new column value (using `NULL` to indicate
 #'   when columns should be deleted), then passes that to `dplyr_col_modify()`.
-#'   `transmute()` does the same then uses 1d `[` to select the columns.
 #'
 #' * `summarise()` works similarly to `mutate()` but the data modified by
 #'   `dplyr_col_modify()` comes from `group_data()`.